requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
azure-core==1.32.0
azure-storage-blob==12.24.1
backports.tarfile==1.2.0
baidu-aip==4.16.13
blinker==1.9.0
blis==0.7.11
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
docutils==0.21.2
elastic-transport==8.17.0
elasticsearch==8.17.0
environs==9.5.0
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.1.0
flatbuffers==25.1.21
fonttools==4.53.1
fsspec==2024.6.1
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.24.6
humanfriendly==10.0
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.4.4
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.8.0
Keras-Preprocessing==1.1.2
keyring==25.6.0
kiwisolver==1.4.5
LAC==2.1.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
libclang==18.1.1
ltp==4.2.14
ltp-core==0.1.4
ltp-extension==0.1.13
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
milvus-model==0.2.12
minio==7.2.15
ml-dtypes==0.3.2
more-itertools==10.6.0
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
networkx==3.2.1
nh3==0.2.21
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opt-einsum==3.3.0
optree==0.14.1
packaging==24.2
paddle-bfloat==0.1.7
paddlepaddle==2.5.0
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
pefile==2023.2.7
pillow==10.4.0
pip==25.0.1
plottable==0.1.5
preshed==3.0.9
prettytable==3.15.1
protobuf==3.20.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.21
pydantic_core==2.27.2
pyecharts==2.0.6
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMySQL==1.1.0
pyparsing==3.1.4
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.2
readme_renderer==44.0
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.4
scikit-learn==1.6.1
scipy==1.13.1
setuptools==75.8.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
SpeechRecognition==3.10.4
srsly==2.5.1
sympy==1.13.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.10.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
thinc==8.1.12
threadpoolctl==3.5.0
tokenizer==3.4.5
tokenizers==0.19.1
tomli==2.2.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
twine==6.1.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zh-core-web-sm==3.5.0
zipp==3.20.1
